Произошел троллинг (рефакторинг всего, начиная с базы данных)

+++ src/main/java/TaskTracker/businessLogic/services/GroupService.java
@@ -3,5 +3,5 @@ package TaskTracker.businessLogic.services;
 import TaskTracker.database.beans.Group;
 
 public interface GroupService {
-    Group getGroup(int id);
+    Group getGroup(Long id);
 }
